requirements.txt
Django==4.2.7
requests==2.31.0
python-decouple==3.8
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-cors-headers==4.3.1
